requirements.txt
# Core ML/DL
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
opencv-python>=4.8.0

# Data Processing & Augmentation
albumentations>=1.3.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Computer Vision Utils
grad-cam>=1.4.8

# ML Utilities
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Model Optimization & Export
onnx>=1.14.0
onnxruntime>=1.15.0

# Experiment Tracking
tensorboard>=2.13.0

# IoT & Communication
paho-mqtt>=1.6.0
requests>=2.31.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0

# Hyperparameter Optimization
optuna>=3.3.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utils
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
pyarrow>=13.0.0
